chore: refactor of jakta examples for paper
